sdk/python/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
pyyaml>=6.0
pytest>=7.0.0
